echo/pyproject.toml
[project]
name = "mcp-server-echo"
version = "0.1.0"
description = "A Model Context Protocol echo server with HTTP transport"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "MCP Echo Server" }]
keywords = ["mcp", "echo", "http", "llm"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "mcp>=1.1.3",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "PyJWT[crypto]>=2.9.0",
    "cryptography==41.0.0",
]

[project.scripts]
mcp-server-echo = "mcp_server_echo:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = ["pyright>=1.1.389", "ruff>=0.7.3"] 